Use 64-bit quantities

+++ driver.c
@@ -1,3 +1,4 @@
+#include <stdint.h>
 #include <stdio.h>
 
 #define fixnum_mask      1
@@ -14,13 +15,14 @@
 
 #define emptylist_tag    0xe
 
-extern int scheme_entry (void);
+extern long scheme_entry (void);
 
 int main (int argc, char **argv)
 {
-  int val = scheme_entry ();
+  unsigned long val = scheme_entry ();
+
   if ((val & fixnum_mask) == fixnum_tag) {
-    printf ("%d\n", (val >> fixnum_shift));
+    printf ("%ld\n", (val >> fixnum_shift));
   } else if ((val & boolean_mask) == boolean_tag) {
     printf ("%s\n", (val >> boolean_shift) ? "#t" : "#f");
   } else if ((val & char_mask) == char_tag) {
@@ -30,5 +32,6 @@ int main (int argc, char **argv)
   } else {
     printf ("uknown type\n");
   }
+
   return 0;
 }
